+++ app-mf-webpack/src/index.tsx
@@ -1,1 +1,5 @@
 import('./bootstarap')
+import React from 'react';
+import ReactRouterDom from 'react-dom';
+window.ReactDOM = ReactRouterDom
+window.React = React;